refactor(loop_detection): graduate to top-level layout, add docs-site page
Merge `main` and relocate from `experimental/loop_detection` to top-level
`pydantic_ai_harness/loop_detection/`, matching #347/#354. Drops the experimental
import warning, fixes import paths, moves tests. Adds `docs/loop-detection.md`
(nav.json, index.md, parity gate) and flips the README Reliability row to Docs.
No behavior change.
